Explicit cast

+++ src/mlpack/methods/nca/nca_main.cpp
@@ -58,7 +58,7 @@ int main(int argc, char* argv[])
   const string outputFile = CLI::GetParam<string>("output_file");
 
   const double stepSize = CLI::GetParam<double>("step_size");
-  const size_t maxIterations = CLI::GetParam<int>("max_iterations");
+  const size_t maxIterations = (size_t) CLI::GetParam<int>("max_iterations");
   const double tolerance = CLI::GetParam<double>("tolerance");
   const bool normalize = CLI::HasParam("normalize");
